+++ src/OrcLib/LocationSet.cpp
@@ -223,9 +223,11 @@ void GetExcludedVolumeLocations(
     const LocationSet::PathExcludes& excludedPaths,
     std::vector<Location::Ptr>& excludedLocations)
 {
+    const bool excludeAll = excludedPaths.find(L"*") != std::cend(excludedPaths);
+
     for (const auto& path : volume.Paths)
     {
-        if (excludedPaths.find(path) != std::cend(excludedPaths))
+        if (excludeAll || excludedPaths.find(path) != std::cend(excludedPaths))
         {
             std::copy(
                 std::cbegin(volume.Locations), std::cend(volume.Locations), std::back_inserter(excludedLocations));
@@ -236,7 +238,7 @@ void GetExcludedVolumeLocations(
 
     for (const auto& location : volume.Locations)
     {
-        if (excludedPaths.find(location->GetLocation()) != std::cend(excludedPaths))
+        if (excludeAll || excludedPaths.find(location->GetLocation()) != std::cend(excludedPaths))
         {
             std::copy(
                 std::cbegin(volume.Locations), std::cend(volume.Locations), std::back_inserter(excludedLocations));
@@ -1092,28 +1094,82 @@ HRESULT LocationSet::AddLocationsFromConfigItem(const ConfigItem& config)
     return hr;
 }
 
-HRESULT LocationSet::AddLocationsFromArgcArgv(int argc, LPCWSTR argv[])
+HRESULT LocationSet::ParseLocationsFromConfigItem(const ConfigItem& config, std::vector<std::wstring>& locations)
 {
-    HRESULT hr = E_FAIL;
+    if (!config)
+    {
+        return S_OK;
+    }
 
-    if (FAILED(hr = EnumerateLocations()))
-        return hr;
+    if (config.Type != ConfigItem::NODELIST)
+    {
+        return E_INVALIDARG;
+    }
 
+    if (config.strName != L"location")
+    {
+        return E_INVALIDARG;
+    }
+
+    HRESULT hr = S_OK;
+    for (auto& item : config.NodeList)
+    {
+        locations.emplace_back(item.c_str());
+    }
+
+    return S_OK;
+}
+
+void LocationSet::ParseLocationsFromArgcArgv(int argc, LPCWSTR argv[], std::vector<std::wstring>& locations)
+{
     for (int i = 1; i < argc; i++)
     {
         if (argv[i][0] != L'/' && argv[i][0] != L'+' && argv[i][0] != L'-')
         {
             if (argv[i][0] == L'*')
             {
-                if (FAILED(hr = ParseAllVolumes()))
-                    return hr;
-                return S_OK;
+                Log::Debug(L"Enable all locations");
             }
-            else
+
+            locations.emplace_back(argv[i]);
+        }
+    }
+}
+
+HRESULT LocationSet::AddLocationsFromArgcArgv(int argc, LPCWSTR argv[])
+{
+    HRESULT hr = E_FAIL;
+
+    if (FAILED(hr = EnumerateLocations()))
+    {
+        return hr;
+    }
+
+    std::vector<std::wstring> locationsFromCli;
+    ParseLocationsFromArgcArgv(argc, argv, locationsFromCli);
+    if (locationsFromCli.empty())
+    {
+        return S_OK;
+    }
+
+    Log::Debug("Disable location set previously as a cli option overrides the value");
+    for (auto& location : m_Locations)
+    {
+        location.second->SetParse(false);
+    }
+
+    for (const auto& location : locationsFromCli)
+    {
+        if (location == L"*")
+        {
+            return ParseAllVolumes();
+        }
+        else
+        {
+            std::vector<std::shared_ptr<Location>> addedLocs;
+            if (FAILED(hr = AddLocations(location.c_str(), addedLocs)))
             {
-                std::vector<std::shared_ptr<Location>> addedLocs;
-                if (FAILED(hr = AddLocations(argv[i], addedLocs)))
-                    return hr;
+                return hr;
             }
         }
     }